modelは切り離した
